ia-cursos-online/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
MarkupSafe==3.0.2
SQLAlchemy==2.0.40
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
cryptography==45.0.3
Pillow==11.2.1
requests==2.32.3
stripe==7.0.0
Flask-Migrate==4.0.5
alembic==1.13.1
